start pulling apart transcription
