+++ src/podcast_processor/podcast_processor.py
@@ -14,9 +14,11 @@ from openai import OpenAI
 from pydub import AudioSegment  # type: ignore[import-untyped]
 
 from .env_settings import populate_env_settings
-
-Segment = Any
-
+from .transcribe import (
+    Segment,
+    Transcriber,
+    RemoteWhisperTranscriber,
+)
 
 env_settings = populate_env_settings()
 
@@ -56,6 +58,7 @@ class PodcastProcessor:
             base_url=env_settings.openai_base_url,
             api_key=env_settings.openai_api_key,
         )
+        self.transcriber: Transcriber = RemoteWhisperTranscriber(self.logger)
 
     def init_pickle_transcripts(self) -> Any:
         pickle_path = "transcripts.pickle"
@@ -166,50 +169,6 @@ class PodcastProcessor:
 
         self.update_pickle_transcripts(task, segments)
         return segments
-
-    def remote_whisper(self, task: PodcastProcessorTask) -> List[Segment]:
-        self.logger.info("Using remote whisper")
-        self.split_file(task.audio_path)
-        all_segments = []
-        for i in range(0, len(os.listdir(f"{task.audio_path}_parts")), 1):
-            segments = self.get_segments_for_chunk(f"{task.audio_path}_parts/{i}.mp3")
-            all_segments.extend(segments)
-        # clean up
-        shutil.rmtree(f"{task.audio_path}_parts")
-        return all_segments
-
-    def get_segments_for_chunk(self, chunk_path: str) -> List[Segment]:
-        with open(chunk_path, "rb") as f:
-            model_extra = self.client.audio.transcriptions.create(
-                model="whisper-1",
-                file=f,
-                timestamp_granularities=["segment"],
-                language="en",
-                response_format="verbose_json",
-            ).model_extra
-
-            assert model_extra is not None
-
-            return model_extra["segments"]
-
-    def split_file(
-        self, audio_path: str, chunk_size_bytes: int = 24 * 1024 * 1024
-    ) -> None:
-        if not os.path.exists(audio_path + "_parts"):
-            os.makedirs(audio_path + "_parts")
-        audio = AudioSegment.from_mp3(audio_path)
-        duration_milliseconds = len(audio)
-        chunk_duration = (
-            chunk_size_bytes / os.path.getsize(audio_path)
-        ) * duration_milliseconds
-        chunk_duration = int(chunk_duration)
-
-        num_chunks = math.ceil(duration_milliseconds / chunk_duration)
-        for i in range(num_chunks):
-            start_time = i * chunk_duration
-            end_time = (i + 1) * chunk_duration
-            chunk = audio[start_time:end_time]
-            chunk.export(f"{audio_path}_parts/{i}.mp3", format="mp3")
 
     def get_user_prompt_template(self, prompt_template_path: str) -> Template:
         with open(prompt_template_path, "r") as f:
